Moved components to designated folders
