+++ src/components/popover/widgets/capture-keystroke/capture-keystroke-button.component.scss
