+++ src/components/popover/widgets/icon/icon.component.scss
