+++ src/components/popover/widgets/icon/icon.component.ts
