+++ src/components/svg/keyboard/svg-keyboard.component.html
