+++ src/components/svg/keyboard/svg-keyboard.component.scss
