+++ src/components/svg/keys/svg-icon-text-key.component.html
