+++ src/components/svg/keys/svg-icon-text-key.component.ts
@@ -14,5 +14,4 @@ export class SvgIconTextKeyComponent implements OnInit {
     constructor() { }
 
     ngOnInit() { }
-
 }
