+++ src/components/svg/keys/svg-keyboard-key.component.html
