+++ src/components/svg/keys/svg-keyboard-key.component.ts
@@ -1,15 +1,15 @@
 import { Component, OnInit, Input, OnChanges, SimpleChange } from '@angular/core';
 import {NgSwitch, NgSwitchCase} from '@angular/common';
 
-import {KeyAction} from '../../../config-serializer/config-items/KeyAction';
-import {KeystrokeAction} from '../../../config-serializer/config-items/KeystrokeAction';
-import {KeyModifiers} from '../../../config-serializer/config-items/KeyModifiers';
-import {PlayMacroAction} from '../../../config-serializer/config-items/PlayMacroAction';
-import {SwitchLayerAction, LayerName}  from '../../../config-serializer/config-items/SwitchLayerAction';
-import {MapperService} from '../../services/mapper.service';
-import {SwitchKeymapAction} from '../../../config-serializer/config-items/SwitchKeymapAction';
-import {UhkConfiguration} from '../../../config-serializer/config-items/UhkConfiguration';
-import {UhkConfigurationService} from '../../services/uhk-configuration.service';
+import {KeyAction} from '../../../../config-serializer/config-items/KeyAction';
+import {KeystrokeAction} from '../../../../config-serializer/config-items/KeystrokeAction';
+import {KeyModifiers} from '../../../../config-serializer/config-items/KeyModifiers';
+import {PlayMacroAction} from '../../../../config-serializer/config-items/PlayMacroAction';
+import {SwitchLayerAction, LayerName}  from '../../../../config-serializer/config-items/SwitchLayerAction';
+import {SwitchKeymapAction} from '../../../../config-serializer/config-items/SwitchKeymapAction';
+import {UhkConfiguration} from '../../../../config-serializer/config-items/UhkConfiguration';
+import {UhkConfigurationService} from '../../../services/uhk-configuration.service';
+import {MapperService} from '../../../services/mapper.service';
 
 import {SvgOneLineTextKeyComponent} from './svg-one-line-text-key.component';
 import {SvgTwoLineTextKeyComponent} from './svg-two-line-text-key.component';
