+++ src/components/svg/keys/svg-keyboard-key.model.ts
