+++ src/components/svg/keys/svg-single-icon-key.component.html
