+++ src/components/svg/keys/svg-single-icon-key.component.ts
@@ -2,17 +2,15 @@ import { Component, OnInit, Input } from '@angular/core';
 
 @Component({
     moduleId: module.id,
-    selector: 'g[svg-text-icon-key]',
-    template: require('./svg-text-icon-key.component.html')
+    selector: 'g[svg-single-icon-key]',
+    template: require('./svg-single-icon-key.component.html')
 })
-export class SvgTextIconKeyComponent implements OnInit {
+export class SvgSingleIconKeyComponent implements OnInit {
     @Input() width: number;
     @Input() height: number;
-    @Input() text: string;
     @Input() icon: string;
 
     constructor() { }
 
     ngOnInit() { }
-
 }
