+++ src/components/svg/keys/svg-switch-keymap-key.component.html
