+++ src/components/svg/keys/svg-switch-keymap-key.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, Input } from '@angular/core';
 
-import {MapperService} from '../../services/mapper.service';
+import {MapperService} from '../../../services/mapper.service';
 
 @Component({
     moduleId: module.id,
@@ -18,5 +18,4 @@ export class SvgSwitchKeymapKeyComponent implements OnInit {
     ngOnInit() {
         this.icon = this.mapperService.getIcon('switch-keymap');
     }
-
 }
