+++ src/components/svg/keys/svg-text-icon-key.component.html
