+++ src/components/svg/module/svg-module.component.scss
