+++ src/components/svg/popover/svg-keyboard-popover.component.scss
