+++ src/components/svg/popover/svg-keyboard-popover.component.ts
@@ -1,9 +1,9 @@
 import { Component, OnInit, Input} from '@angular/core';
 
-import {Module} from '../../config-serializer/config-items/Module';
-import {KeyAction} from '../../config-serializer/config-items/KeyAction';
-import {SvgKeyboardComponent} from './svg-keyboard.component';
-import {PopoverComponent} from './popover/popover.component';
+import {Module} from '../../../../config-serializer/config-items/Module';
+import {KeyAction} from '../../../../config-serializer/config-items/KeyAction';
+import {SvgKeyboardComponent} from '../keyboard/svg-keyboard.component';
+import {PopoverComponent} from '../../popover/popover.component';
 
 @Component({
     selector: 'svg-keyboard-popover',
